refactor run map code to be used in resuts
