+++ src/App.js
@@ -14,6 +14,7 @@ import RunsMap from "./Components/RunsMap/RunsMap";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import { mdiProgressUpload } from "@mdi/js";
+import Results from "./Components/Results/Results";
 const App = () => {
   const [userData, setUserData] = useState(null);
   const [loading, setLoading] = useState(false); //passed down to other components as props, a boolean which toggles the loading spinner overlay
@@ -96,6 +97,14 @@ const App = () => {
               <PrivateRoute
                 path="/runs"
                 component={RunsMap}
+                userData={userData}
+                fetchData={fetchData}
+                setLoading={setLoading}
+                toast={toast}
+              />
+              <PrivateRoute
+                path="/results"
+                component={Results}
                 userData={userData}
                 fetchData={fetchData}
                 setLoading={setLoading}
