+++ src/Components/Results/Results.js
@@ -1,0 +1,11 @@
+import React, { useEffect } from "react";
+
+const Results = (props) => {
+  return (
+    <div>
+      <h1>Results</h1>
+    </div>
+  );
+};
+
+export default Results;
