+++ src/Components/RunsMap/RunsMap.js
@@ -8,8 +8,8 @@ import Player from "../Player/Player";
 import BurgerMenu from "../BurgerMenu/BurgerMenu";
 import { useHistory } from "react-router-dom";
 import mapboxgl from "mapbox-gl/dist/mapbox-gl";
-import { isProduction } from "../../Common/CommonFunctions";
 
+import { retrieveRuns } from "../../Functions/RetrieveRuns";
 // @ts-ignore
 // eslint-disable-next-line import/no-webpack-loader-syntax
 mapboxgl.workerClass = require("worker-loader!mapbox-gl/dist/mapbox-gl-csp-worker").default;
@@ -130,78 +130,30 @@ const RunsMap = (props) => {
       stopRun();
     };
   }, []);
+
+  const setStates = (vals) => {
+    setRunIdList(vals[0]);
+    setRunList(vals[1]);
+    spotifyToken = vals[2];
+    let runData = vals[1][vals[1].length - 1];
+    console.log(runData);
+    setRun(runData);
+  };
+
   useEffect(async () => {
-    window.addEventListener("resize", function () {
-      setWindowWidth(window.innerWidth);
-    });
-    props.setLoading(true);
     if (props.userData == null) {
       await props.fetchData(firebaseApp.auth().currentUser.uid);
 
       return;
     }
-    console.log(props.userData.spotifyRefreshToken);
-    let uri = isProduction()
-      ? "https://europe-west2-musicmakesyourunfaster.cloudfunctions.net/app/api/spotify/refresh"
-      : "http://localhost:5000/musicmakesyourunfaster/europe-west2/app/api/spotify/refresh";
-    spotifyToken = await fetch(uri, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        refresh_token: props.userData.spotifyRefreshToken,
-      }),
+
+    window.addEventListener("resize", function () {
+      setWindowWidth(window.innerWidth);
     });
-    spotifyToken = await spotifyToken.text();
-    console.log(spotifyToken);
 
-    let runRef = await db
-      .collection("users")
-      .doc(props.userData.fitbitId)
-      .collection("runs")
-      .get();
-    console.log(runRef.docs);
-    if (runRef.empty) {
-      props.toast.error("❗ No runs found, please pull a recent run first. ❗");
-      props.setLoading(false);
-      history.push("/dashboard");
-      return;
-    }
+    await retrieveRuns(props, setStates);
 
-    let idList = [];
-    for (let i = 0; i < runRef.docs.length; i++) {
-      console.log(runRef.docs[i].id);
-      if (runRef.docs[i].id.includes("part")) {
-        if (!runRef.docs[i].id.includes("part 0")) {
-          //only use one part
-          continue;
-        }
-      }
-      idList.push(runRef.docs[i].id);
-    }
-    let allRuns = [];
-
-    //Add runs to list, if split into subparts, combine them.
-    let fix_index = 0; //move this back one when we encounter any id with 'part' in it
-    runRef.docs.forEach((r, index) => {
-      if (r.id.includes("part")) {
-        console.log("0 index ", fix_index);
-        if (!r.id.includes("part 0")) {
-          let add_here = fix_index - parseInt(r.id.split(" ")[2]);
-          allRuns[add_here] = [...allRuns[add_here], ...r.data().run_map];
-          return;
-        }
-      }
-      allRuns.push(r.data().run_map);
-      fix_index++;
-    });
-    setRunIdList(idList);
-    setRunList(allRuns);
-    let runData = allRuns[allRuns.length - 1];
-    setRun(runData);
     replayCounter = 0;
-    props.setLoading(false);
   }, [props.userData]);
 
   if (run == undefined) {
@@ -356,7 +308,11 @@ const RunsMap = (props) => {
             marginTop: "2rem",
           }}
         >
-          <Player uri={currentlyPlaying.uri} token={spotifyToken} play={play} />
+          <Player
+            uri={[currentlyPlaying.uri]}
+            token={spotifyToken}
+            play={play}
+          />
         </div>
       ) : null}
     </div>
